before: widen the route DP's cost components to u64 with a sub-sentinel saturation ceiling
A feasible expansion chain's distance saturated at u32::MAX, which is
also the infeasible sentinel: at 2^32 id levels (a ~1 GiB packed id) the
chain read as infeasible, the tie resolved right into an absent sibling,
and the splice emit panicked in debug (id-cursor desync in release).
Cost components are u64 now, and feasible folds saturate at
Cost::CEILING, strictly below Cost::INFEASIBLE, so no chain of any
encodable depth can alias into the sentinel: the corner closes by
construction. The recursive oracle's grow and the reference route probe
step components through the same Cost::deepen, replacing the oracle's
unchecked '+ 1' (a debug overflow past the old bound), so every
implementation of the fold agrees by construction; the brute-force
enumeration stays an independent witness (structural Option
infeasibility, exact arithmetic over bounded test trees).

The rise loop's saturation bound is a named parameter (production
passes Cost::CEILING), and a scaled-sentinel family pins the contract
in constructible range: direction chains of arbitrary depth and
orientation, crossed with ceilings they can reach and pass, must fold
to a strictly feasible saturated distance with every route bit turned
into the present child - plus the deterministic one-past-the-bound
witness at ceiling 7 and its end-to-end tick differential. Both read
red under alias-into-sentinel semantics (verified by reversible
mutation, restored).

Meter tick rows measured at the parent and after: every counter column
byte-identical (the route DP's transients are width-priced bit stacks
and the route's bit vector, none of which the widening moves); no pin
moved.

+++ crates/before/src/testing/grow_brute_force.rs
@@ -29,8 +29,13 @@ use std::sync::Arc;
 use crate::oracle;
 
 /// The inflation cost the paper assigns: `(expansions, depth)`, lexicographic.
-/// Matches the oracle's `Cost` and the impl's `grow::Cost`.
-pub(crate) type GrowCost = (u32, u32);
+///
+/// Matches the oracle's `Cost` and the impl's `grow::Cost` component width.
+/// Deliberately unchecked exact arithmetic, no saturation and no infeasible
+/// sentinel: infeasibility is structural here (an empty enumeration /
+/// [`None`]), and depths are bounded by the enumerated test trees, so the
+/// brute force stays an independent witness of the DPs' saturating folds.
+pub(crate) type GrowCost = (u64, u64);
 
 /// Every feasible single-region inflation of `(id, e)`, each paired with its true
 /// `(expansions, depth)` cost.
